# Make IOFS.ReadDir check for fs.ReadDirFile